Adding first and last name validation, separated my code from Venus'
