+++ index.css
@@ -27,23 +27,18 @@ font-size: 50px;
 	border-color: #777;
 }
 
- /* Input Messages */
+ /* User-side Messages */
 .form small {
-<<<<<<< HEAD
-	color: #000;
-	background-color: #efefef;
-=======
-	color: #e74c3c;
-	/* background-color: #efefef; */
->>>>>>> 7bf90aee791e15ca73e1733820db24e641f29e48
 	position: absolute;
 	visibility: hidden;
 }
 
 /* Error Message */
 .form.error small {
-	color: red;
-	background-color: efefef;
+	margin-top: .2em;
+	margin-left: 1em;
+	background-color: red;
+	color: white;
 	visibility: visible;
 }
 
@@ -53,12 +48,19 @@ font-size: 50px;
 
 /* Success Message */
 .form.success small {
-	color: #2ecc71;
-	background-color: #efefef;
+	margin-top: .2em;
+	margin-left: 1em;
+	background-color: #2ecc71;
+	color: white;
 	visibility: visible;
 }
 
 .form.success input{
 	border-color: #2ecc71;
+}
+
+i {
+	visibility: hidden;
+	position: absolute;
 }
