+++ index.js
@@ -1,4 +1,4 @@
-//Venus
+/*//Venus
 // Validate first / last name / nothings red but can not tell if works//;
 
       //Input fields
@@ -96,69 +96,6 @@
           return false;
       }  
   
-      }
-  //------------------------------
-
-  //Lexie
-//Validate Phone Number
-
-//Get info from html
-const form = document.getElementById('form');
-const phoneNumber = document.getElementById('phonenumber');
-
-//Button Event Listener Function
-form.addEventListener('submit' , (e) => {
-    e.preventDefault();
-       
-    //Validate input
-    checkInputs();
-
-});
-
-//Check phone number input
-function checkInputs() {
-    const phoneNumberValue = phoneNumber.value.trim();
-
-    //Validation Errors & Success
-    if(phoneNumberValue == "") {
-        setErrorFor(phoneNumber, 'Phone number cannot be blank');
-    } else if (!validatePhone(phoneNumberValue)){
-        setErrorFor(phoneNumber, 'Phone number is not valid');
-        console.log(phoneNumberValue + " not valid");
-    } else if (validatePhone(phoneNumberValue)) {
-        setSuccessFor(phoneNumber, 'Success!');
-        console.log(phoneNumberValue);
-        form.reset();
-    }
-
-//display error message in RED to user
-function setErrorFor(input, message) {
-    const formControl = input.parentElement;
-    const small = formControl.querySelector('small');
-
-    //display error message user side
-    small.innerText = message;
-
-    //add error class to style in css
-    formControl.className = 'form error';
-}
-
-//display success message to user
-//GREEN added in CSS
-function setSuccessFor(input, message) {
-    const formControl = input.parentElement;
-    const small = formControl.querySelector('small');
-
-    //display success message user side
-    small.innerText = message;
-
-    //add success class for styling in css
-    formControl.className = 'form success';
-}
-
-function validatePhone(phoneNumber) {
-    //regex
-    return /^(?:\+?1[-. ]?)?\(?([0-9]{3})\)?[-. ]?([0-9]{3})[-. ]?([0-9]{4})$/.test(phoneNumber);
-}}
-// }
-;
+      }}
+      */
+
